Fix HTML tag removal with improved regex handling of incomplete tags
